Move all commands to Query\Command namespace
