+++ src/QueryType/ManagedResources/Query/Command/Synonyms/Create.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Solarium\QueryType\ManagedResources\Query\Stopwords\Command;
+namespace Solarium\QueryType\ManagedResources\Query\Command\Synonyms;
 
 use Solarium\QueryType\ManagedResources\Query\Command\AbstractCreate;
 
@@ -11,6 +11,6 @@ class Create extends AbstractCreate
      */
     public function getRawData(): string
     {
-        return json_encode(['class' => 'org.apache.solr.rest.schema.analysis.ManagedWordSetResource']);
+        return json_encode(['class' => 'org.apache.solr.rest.schema.analysis.ManagedSynonymGraphFilterFactory$SynonymManager']);
     }
 }
