+++ src/QueryType/ManagedResources/Query/Stopwords.php
@@ -8,8 +8,8 @@ use Solarium\QueryType\ManagedResources\Query\Command\Config;
 use Solarium\QueryType\ManagedResources\Query\Command\Delete;
 use Solarium\QueryType\ManagedResources\Query\Command\Exists;
 use Solarium\QueryType\ManagedResources\Query\Command\Remove;
-use Solarium\QueryType\ManagedResources\Query\Stopwords\Command\Add;
-use Solarium\QueryType\ManagedResources\Query\Stopwords\Command\Create;
+use Solarium\QueryType\ManagedResources\Query\Command\Stopwords\Add;
+use Solarium\QueryType\ManagedResources\Query\Command\Stopwords\Create;
 use Solarium\QueryType\ManagedResources\Query\Stopwords\InitArgs;
 use Solarium\QueryType\ManagedResources\ResponseParser\Stopwords as ResponseParser;
 use Solarium\QueryType\ManagedResources\Result\Stopwords\WordSet;
