+++ src/QueryType/ManagedResources/Query/Synonyms.php
@@ -8,8 +8,8 @@ use Solarium\QueryType\ManagedResources\Query\Command\Config;
 use Solarium\QueryType\ManagedResources\Query\Command\Delete;
 use Solarium\QueryType\ManagedResources\Query\Command\Exists;
 use Solarium\QueryType\ManagedResources\Query\Command\Remove;
-use Solarium\QueryType\ManagedResources\Query\Synonyms\Command\Add;
-use Solarium\QueryType\ManagedResources\Query\Synonyms\Command\Create;
+use Solarium\QueryType\ManagedResources\Query\Command\Synonyms\Add;
+use Solarium\QueryType\ManagedResources\Query\Command\Synonyms\Create;
 use Solarium\QueryType\ManagedResources\Query\Synonyms\InitArgs;
 use Solarium\QueryType\ManagedResources\ResponseParser\Synonyms as ResponseParser;
 use Solarium\QueryType\ManagedResources\Result\Synonyms\SynonymMappings;
