+++ tests/QueryType/ManagedResources/RequestBuilder/SynonymsTest.php
@@ -11,9 +11,9 @@ use Solarium\QueryType\ManagedResources\Query\Command\Config as ConfigCommand;
 use Solarium\QueryType\ManagedResources\Query\Command\Delete as DeleteCommand;
 use Solarium\QueryType\ManagedResources\Query\Command\Exists as ExistsCommand;
 use Solarium\QueryType\ManagedResources\Query\Command\Remove as RemoveCommand;
+use Solarium\QueryType\ManagedResources\Query\Command\Synonyms\Add as AddCommand;
+use Solarium\QueryType\ManagedResources\Query\Command\Synonyms\Create as CreateCommand;
 use Solarium\QueryType\ManagedResources\Query\Synonyms as SynonymsQuery;
-use Solarium\QueryType\ManagedResources\Query\Synonyms\Command\Add as AddCommand;
-use Solarium\QueryType\ManagedResources\Query\Synonyms\Command\Create as CreateCommand;
 use Solarium\QueryType\ManagedResources\Query\Synonyms\InitArgs;
 use Solarium\QueryType\ManagedResources\RequestBuilder\Resource as SynonymsRequestBuilder;
 use Solarium\Tests\Integration\TestClientFactory;
